Toggle building examples and fixed Windows issues
- Added MATPLOTLIBCPP_BUILD_EXAMPLES to CMake (good for disabling when importing this into other projects)
- Fixed several cast warnings
- Fixed Python header trying to link to debug library

+++ examples/colorbar.cpp
@@ -13,7 +13,7 @@ int main()
     std::vector<float> z(ncols * nrows);
     for (int j=0; j<nrows; ++j) {
         for (int i=0; i<ncols; ++i) {
-            z.at(ncols * j + i) = std::sin(std::hypot(i - ncols/2, j - nrows/2));
+            z.at(ncols * j + i) = static_cast<float>(std::sin(std::hypot(i - ncols/2, j - nrows/2)));
         }
     }
